backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
psycopg2==2.9.1
django-heroku==0.3.1
django-environ==0.8.1
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2